projects/publisher/pyproject.toml
[tool.poetry]
name = "publisher"
version = "0.1.0"
description = ""
authors = ["Pavel Pantiukhov <pantyukhov.p@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.7"
dj-database-url = "^1.2.0"
environs = "^9.5.0"
django-redis = "^5.2.0"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"
djangorestframework-dataclasses = "^1.2.0"
xrpl-py = "^1.7.0"
wallet = "^2.1.525"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0", "pip>=20.0.0"]
build-backend = "poetry.core.masonry.api"
